etl/requirements.txt
azure-storage-blob==12.19.0
pandas==2.1.4
pyodbc==5.0.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
pytest==7.4.4
pytest-cov==4.1.0
